Moves aiosqlite/persister import to plugins
